Add project meta data to the Subscription
Previously the project id was added to the Project line item on
the subscription not the subscription for all but the fist project

While stored it's not visible in the dashboard.

Added code to add all projects to the subscriptions metadata.

We might want to revert this when/if we have projects at different
price points.

Fixes #423