git: Ignore generated files
Artifacts that are generated through assets in the repository, should
not be part of the repository. This is the point of feature like
Releases at GitHub for example.

Generated assets increase the repository size, create noisy diffs, and
merge conflicts. As such, this change removes the dist directory from
the repository going forward.