TextInput: set type over password flag
Before this change TextInput components had a boolean field to set to
true for password fields. This change removes that field in favour of a
'type' field. This field can now be set to "text" (default) or
"password".

Note: this is a breaking change, and requires changes in dependant
projects.