Limit password length to 1024 chars
The fastify body parser already limits the body of any post to 1mb
this just makes sure we don't hit that with a stupidly long password.

We can bcrypt hash a 1mb payload in aproxy 102ms with nodejs so
shouldn't be a problem, but 1024 seams like an excsively long password.

fixes #921