docs: Deduplicate Mosquitto install docs
Since Deployments/Devices work without MQTT there's not a pressing need
to setup Mosquitto. Given it's the hardest part of the whole process,
I've moved down the first mention of it, to the point where the user has
mostly set up FlowForge to the point where it's functioning and they're
successful.