Disable filestore by default
As this is a licensed feature it is already diabled by default,
but the localfs installer does not include the file-server. So
When you add a license to the localfs build it enables
persistent context and fails to start instance.

With this off by default it will require file-server to be
installed and started and configured.

Will look at bundling file-server in installer, but this is best
safe quick option