Add more context to the snapshot selector in the pipeline stages
Now that we have the combobox available, we can type search through options.

I extended the combobox to be able to take on multiple search parameters, even on nested keys and also allowed for the customization of its options list.

I swapped the listbox on the pipelines stage selector with the combobox, added the user and description for the user to see when selecting options while also extending the search functionality to search through the username and description.