fix(expert): make togglePinWithWidth inject optional
The Expert component was crashing in the immersive editor when sending
messages because it expected togglePinWithWidth to be provided by a
parent component. In the immersive editor, Expert is loaded directly
as a route component without the RightDrawer wrapper that provides
this function.

Changed the inject from array to object format with a default no-op
function, allowing the component to work both in the drawer context
and standalone in the immersive editor.