Allow default TeamType to enabled all feats default
fixes #4866

Adds a check box to the TeamType config that enables all features.

This defaults to enabled for the default team that is create on
first startup.

Need to decide if we want a migration to enable this for existing
installs if there is only 1 team type.

This PR is built on top of #6247 and targets it (so it should be merged
first).